requirements.txt
django==2.1.*
djangorestframework==3.11.0
requests
pillow
